requirements.txt
django
numpy
django-import-export
xhtml2pdf